Refactor and fix lint errors
