+++ packages/client/src/ui/validators/CeloValidators.tsx
@@ -55,7 +55,6 @@ import {
   PageScrollableContent,
   View,
 } from "ui/SharedComponents";
-import Toast from "ui/Toast";
 import {
   RowItem,
   RowItemHeader,
